requirements.txt
fastapi
uvicorn
langchain_core
cdp_langchain
langchain_openai
langgraph
python-dotenv
typing
pydantic